requirements.txt
termcolor
ipaddress
pyyaml
validatorss